+++ frontend/public/locales/es/translation.json
@@ -68,6 +68,8 @@
   "Builds an Authorization Policy based on Rules": "Builds an Authorization Policy based on Rules",
   "CA Certificates": "CA Certificates",
   "Cancel": "Cancel",
+  "Check Config": "Check Config",
+  "Check for possible configuration(s) of external_services.tracing": "Check for possible configuration(s) of external_services.tracing",
   "Check Status": "Check Status",
   "Check the usual ports for the tracing service and provide a subset of the tracing configuration based on the tracing services found for external_services.tracing.": "Check the usual ports for the tracing service and provide a subset of the tracing configuration based on the tracing services found for external_services.tracing.",
   "Choose the traffic rates used to generate the graph. Each supported protocol offers one or more options. Unused protocols can be omitted.": "Choose the traffic rates used to generate the graph. Each supported protocol offers one or more options. Unused protocols can be omitted.",
@@ -173,6 +175,7 @@
   "Every 5m": "Cada 5m",
   "Expand view": "Vista expandida",
   "Extension": "Extension",
+  "external_services.tracing configuration:": "external_services.tracing configuration:",
   "Failed Request": "Failed Request",
   "Failure": "Failure",
   "Fault Injection": "Fault Injection",
@@ -351,7 +354,6 @@
   "Pod Status": "Pod Status",
   "Policy": "Policy",
   "Port": "Port",
-  "Possible configuration(s) found for external_services.tracing": "Possible configuration(s) found for external_services.tracing",
   "Present": "Present",
   "Preview": "Preview",
   "Preview new istio objects": "Preview new istio objects",
@@ -422,6 +424,7 @@
   "TCP Connection": "TCP Connection",
   "TCP Traffic": "TCP Traffic",
   "TCP Traffic Shifting": "TCP Traffic Shifting",
+  "Test Configuration": "Test Configuration",
   "Test tracing config": "Test tracing config",
   "The control plane must sync each proxy when there is a configuration change. This is the average time spent, per proxy,                  to perform the update. N/A means there was no syncing performed for the configured time period.": "The control plane must sync each proxy when there is a configuration change. This is the average time spent, per proxy,                  to perform the update. N/A means there was no syncing performed for the configured time period.",
   "The destination hosts to which traffic is being sent. Enter one or multiple hosts separated by comma.": "The destination hosts to which traffic is being sent. Enter one or multiple hosts separated by comma.",
@@ -453,6 +456,7 @@
   "To": "To",
   "To List": "To List",
   "To read the FAQ entry about why your graph may be empty, ": "To read the FAQ entry about why your graph may be empty, ",
+  "Tracing configuration": "Tracing configuration",
   "Tracing Services Check Logs": "Tracing Services Check Logs",
   "Traffic": "Tráfico",
   "Traffic Animation": "Traffic Animation",
@@ -486,7 +490,6 @@
   "Version: {{version}}": "Versión: {{version}}",
   "View Annotations": "View Annotations",
   "View in Tracing": "View in Tracing",
-  "View logs": "View logs",
   "View on Graph": "View on Graph",
   "Virtual Service / Request Routing": "Virtual Service / Request Routing",
   "VirtualService Host '*' wildcard not allowed on mesh gateway.": "VirtualService Host '*' wildcard not allowed on mesh gateway.",
