+++ frontend/public/locales/zh/translation.json
@@ -58,6 +58,8 @@
   "Builds an Authorization Policy based on Rules": "Builds an Authorization Policy based on Rules",
   "CA Certificates": "CA证书",
   "Cancel": "取消",
+  "Check Config": "Check Config",
+  "Check for possible configuration(s) of external_services.tracing": "Check for possible configuration(s) of external_services.tracing",
   "Check Status": "Check Status",
   "Check the usual ports for the tracing service and provide a subset of the tracing configuration based on the tracing services found for external_services.tracing.": "Check the usual ports for the tracing service and provide a subset of the tracing configuration based on the tracing services found for external_services.tracing.",
   "Choose the traffic rates used to generate the graph. Each supported protocol offers one or more options. Unused protocols can be omitted.": "选择用于生成图表的流量速率。每个支持的协议都提供一个或多个选项。未使用的协议可以省略。",
@@ -161,6 +163,7 @@
   "Every 5m": "每 5m",
   "Expand view": "展开视图",
   "Extension": "Extension",
+  "external_services.tracing configuration:": "external_services.tracing configuration:",
   "Failed Request": "失败的请求",
   "Failure": "Failure",
   "Fault Injection": "故障注入",
@@ -339,7 +342,6 @@
   "Pod Status": "Pod Status",
   "Policy": "Policy",
   "Port": "端口",
-  "Possible configuration(s) found for external_services.tracing": "Possible configuration(s) found for external_services.tracing",
   "Present": "Present",
   "Preview": "预览",
   "Preview new istio objects": "Preview new istio objects",
@@ -410,6 +412,7 @@
   "TCP Connection": "TCP Connection",
   "TCP Traffic": "TCP Traffic",
   "TCP Traffic Shifting": "TCP 流量调整",
+  "Test Configuration": "Test Configuration",
   "Test tracing config": "Test tracing config",
   "The control plane must sync each proxy when there is a configuration change. This is the average time spent, per proxy,                  to perform the update. N/A means there was no syncing performed for the configured time period.": "The control plane must sync each proxy when there is a configuration change. This is the average time spent, per proxy,                  to perform the update. N/A means there was no syncing performed for the configured time period.",
   "The destination hosts to which traffic is being sent. Enter one or multiple hosts separated by comma.": "发送流量时的目标主机，输入一个或多个主机，以逗号分隔。",
@@ -441,6 +444,7 @@
   "To": "To",
   "To List": "To List",
   "To read the FAQ entry about why your graph may be empty, ": "To read the FAQ entry about why your graph may be empty, ",
+  "Tracing configuration": "Tracing configuration",
   "Tracing Services Check Logs": "Tracing Services Check Logs",
   "Traffic": "流量",
   "Traffic Animation": "Traffic Animation",
@@ -474,7 +478,6 @@
   "Version: {{version}}": "Version: {{version}}",
   "View Annotations": "查看Annotations",
   "View in Tracing": "View in Tracing",
-  "View logs": "View logs",
   "View on Graph": "View on Graph",
   "Virtual Service / Request Routing": "Virtual Service / Request Routing",
   "VirtualService Host '*' wildcard not allowed on mesh gateway.": "网状网关上不允许虚拟服务主机'*'通配符。",
